SSPO/requirements.txt
# python >= 3.10.0
torch==2.5.1 # install torch first
accelerate==1.0.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alpaca_eval==0.6.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
av==14.0.1
bitsandbytes==0.45.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
datasets==3.1.0
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.0
fastapi==0.115.6
ffmpy==0.5.0
filelock==3.16.1
fire==0.7.0
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.12
GitPython==3.1.44
gradio==5.12.0
gradio_client==1.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
jieba==0.42.1
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
kiwisolver==1.4.8
-e git+https://github.com/hiyouga/LLaMA-Factory.git@de9bc3fefa4fcb5db7d04589b16282a078c62cb2#egg=llamafactory
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
openai==1.60.2
orjson==3.10.15
packaging==24.2
pandas==2.2.3
patsy==1.0.1
peft==0.12.0
pillow==11.1.0
platformdirs==4.3.6
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
pyarrow==19.0.0
pydantic==2.10.5
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rouge-chinese==1.0.3
ruff==0.9.2
safehttpx==0.1.6
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.20.0
setproctitle==1.3.4
shellingham==1.5.4
shtab==1.7.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.41.3
sympy==1.13.1
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.3
tomlkit==0.13.2
tqdm==4.67.1
transformers==4.46.1
triton==3.1.0
trl==0.9.6
typer==0.15.1
typing_extensions==4.12.2
tyro==0.8.14
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
wandb==0.19.5
websockets==14.2
xxhash==3.5.0
yarl==1.18.3
